requirements.txt
Flask==2.1.0
numpy==1.22.4
pandas==1.3.4
pymongo==4.1.1
python-dotenv==0.20.0
scikit_learn==1.1.1
tqdm==4.62.3
pymongo[srv]==4.1.1
gunicorn
flask-cors == 3.8.0